Skip removing .vscode directory
When copying available files to the .backup directory, empty directories are removed after the move operation.
With this change, the .vscode directory is excluded from deletion.

+++ workspaces/mi/mi-extension/src/util/migrationUtils.ts
@@ -2131,7 +2131,9 @@ function moveFiles(sourcePath: string, destinationPath: string, depth: number = 
 
         if (isDirectory) {
             moveFiles(sourceItemPath, destinationItemPath, depth + 1);
-            fs.rmSync(sourceItemPath, { recursive: true });
+            if (item !== ".vscode") {
+                fs.rmSync(sourceItemPath, { recursive: true });
+            }
         } else {
             fs.renameSync(sourceItemPath, destinationItemPath);
         }
